refactor(auth): rename repository to gateway and introduce service
