+++ eslint.config.mjs
@@ -52,7 +52,7 @@ export default [
       'no-restricted-imports': [
         'error',
         {
-          patterns: ['../*', './/*'],
+          patterns: ['../*', './*'],
           paths: [
             {
               name: 'zod',
